pyproject.toml
[project]
name = "dota2-auto-accept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "pytest>=8.4.1",
]
